requirements.txt
streamlit
groq
python-dotenv
pytest
hypothesis

# Phase 1 — REST API
flask>=3.0,<4
pydantic>=2.6,<3
gunicorn>=21.2,<23

# Phase 2 — Persistence
sqlalchemy>=2.0,<3
alembic>=1.13,<2
psycopg[binary]>=3.1,<4

# Phase 3 — Auth
argon2-cffi>=23.1,<25
PyJWT>=2.8,<3
flask-limiter>=3.5,<4
flask-cors>=4.0,<5
email-validator>=2.1,<3
